// Aadnanmt - Generator Prompt Veo 3/types.ts
export interface FormData {
  subjek: string;
  aksi: string;
  ekspresi: string;
  tempat: string;
  waktu: string;
  gerakanKamera: string;
  pencahayaan: string;
  gayaVideo: string;
  suasanaVideo: string;
  suaraMusik: string;
  kalimatYangDiucapkan: string;
  detailTambahan: string;
}

export interface Option {
  value: string;
  label: string;
}